workers/bot_mock/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
pydantic-settings>=2.2.1
requests>=2.31.0
colorlog>=6.7.0
python-dotenv>=1.0.0
backoff>=2.2.1
PyYAML>=6.0.1
colorama>=0.4.6
tqdm>=4.66.1
tenacity>=8.2.3
werkzeug==2.3.7
waitress==2.1.2
faker==18.13.0
pytz==2023.3 